feat: seperated wasm specific functionality to separate files and do conditional compilation with build tags
