+++ internal/receiver/receiver.go
@@ -5,9 +5,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"net"
-	"net/http"
-	"time"
 
 	"github.com/SpatiumPortae/portal/internal/conn"
 	"github.com/SpatiumPortae/portal/internal/password"
@@ -104,76 +101,7 @@ func Receive(tc conn.Transfer, dst io.Writer, msgs ...chan interface{}) error {
 	if len(msgs) > 0 {
 		msgs[0] <- msg.Payload.PayloadSize
 	}
-	return receive(tc, net.TCPAddr{IP: msg.Payload.IP, Port: msg.Payload.Port}, dst, msgs...)
-}
-
-// receive performs the transfer protocol on the receiving end.
-func receive(relay conn.Transfer, addr net.TCPAddr, dst io.Writer, msgs ...chan interface{}) error {
-
-	// Retrieve a unencrypted channel to rendezvous.
-	rc := conn.Rendezvous{Conn: relay.Conn}
-	// Determine if we should do direct or relay transfer.
-	var tc conn.Transfer
-	direct, err := probeSender(addr, relay.Key())
-	if err != nil {
-		tc = relay
-		// Communicate to the sender that we are using relay transfer.
-		if err := relay.WriteMsg(transfer.Msg{Type: transfer.ReceiverRelayCommunication}); err != nil {
-			return err
-		}
-		_, err := relay.ReadMsg(transfer.SenderRelayAck)
-		if err != nil {
-			return err
-		}
-
-		if len(msgs) > 0 {
-			msgs[0] <- transfer.Relay
-		}
-	} else {
-		tc = direct
-		// Communicate to the sender that we are doing direct communication.
-		if err := relay.WriteMsg(transfer.Msg{Type: transfer.ReceiverDirectCommunication}); err != nil {
-			return err
-		}
-
-		// Tell rendezvous server that we can close the connection.
-		if err := rc.WriteMsg(rendezvous.Msg{Type: rendezvous.ReceiverToRendezvousClose}); err != nil {
-			return err
-		}
-
-		if len(msgs) > 0 {
-			msgs[0] <- transfer.Direct
-		}
-	}
-
-	// Request the payload and receive it.
-	if tc.WriteMsg(transfer.Msg{Type: transfer.ReceiverRequestPayload}) != nil {
-		return err
-	}
-	if err := receivePayload(tc, dst, msgs...); err != nil {
-		return err
-	}
-
-	// Closing handshake.
-	if err := tc.WriteMsg(transfer.Msg{Type: transfer.ReceiverPayloadAck}); err != nil {
-		return err
-	}
-
-	_, err = tc.ReadMsg(transfer.SenderClosing)
-
-	if err != nil {
-		return err
-	}
-
-	if err := tc.WriteMsg(transfer.Msg{Type: transfer.ReceiverClosingAck}); err != nil {
-		return err
-	}
-
-	// Tell rendezvous to close connection.
-	if err := rc.WriteMsg(rendezvous.Msg{Type: rendezvous.ReceiverToRendezvousClose}); err != nil {
-		return err
-	}
-	return nil
+	return doReceive(tc, fmt.Sprintf("%s:%d", msg.Payload.IP, msg.Payload.Port), dst, msgs...)
 }
 
 // receivePayload receives the payload over the provided connection and writes it into the desired location.
@@ -203,30 +131,4 @@ func receivePayload(tc conn.Transfer, dst io.Writer, msgs ...chan interface{}) e
 		}
 	}
 	return nil
-}
-
-// probeSender will try to connect directly to the sender using a linear back off for up to 3 seconds.
-// Returns a transfer connection channel if it succeeds, otherwise it returns an error.
-func probeSender(addr net.TCPAddr, key []byte) (conn.Transfer, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second) // wait at most 3 seconds.
-	defer cancel()
-	d := 250 * time.Millisecond
-	for {
-		select {
-		case <-ctx.Done():
-			return conn.Transfer{}, fmt.Errorf("could not establish a connection to the sender server")
-
-		default:
-			ws, _, err := websocket.Dial(
-				context.Background(), fmt.Sprintf("ws://%s/portal", addr.String()),
-				&websocket.DialOptions{HTTPClient: &http.Client{Timeout: d}},
-			)
-			if err != nil {
-				time.Sleep(d)
-				d = d * 2
-				continue
-			}
-			return conn.TransferFromKey(&conn.WS{Conn: ws}, key), nil
-		}
-	}
 }
